implementation service of the forecast

+++ app/src/main/java/com/example/wecli/data/forecastData/ForecastMain.kt
@@ -12,5 +12,5 @@ data class ForecastMain(
     val sea_level: Int,
     val grnd_level: Int,
     val humidity: Int,
-    val temp_kf: Int
+    val temp_kf: Double
 )
